// MiniHotel.Client/src/app/api/models/booking-create-by-reception-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PaymentMethod } from '../models/payment-method';
export interface BookingCreateByReceptionDto {
  endDate: string;
  fullName: string;
  paymentMethod: PaymentMethod;
  phoneNumber: string;
  roomNumber: string;
  startDate: string;
}